% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualtricsAPI.R
\name{aggregate_account_contacts}
\alias{aggregate_account_contacts}
\title{aggregate_account_contacts}
\usage{
aggregate_account_contacts(name_pattern_regex = ".*")
}
\arguments{
\item{name_pattern_regex}{By default, the function returns contacts from any mailing list with a name matching ".*" ie. all lists. Enter a regex string to match if you want to only aggregate contacts from certain lists.}
}
\value{
The function returns a data.frame
}
\description{
Function to aggregate contacts from all mailing lists in your account.
}
