% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualtricsAPI.R
\name{get_contacts}
\alias{get_contacts}
\title{get_contacts}
\usage{
get_contacts(mailing_list_id)
}
\arguments{
\item{mailing_list_id}{}
}
\value{
The function returns a data.frame
}
\description{
Function to simplify getting contacts from a mailing list.
}
