% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualtricsAPI.R
\name{QualtricsAPI}
\alias{QualtricsAPI}
\title{QualtricsAPI}
\usage{
QualtricsAPI(
  Request,
  QualBaseUrl,
  Qualtrics_Username,
  Clear_Token = FALSE,
  Method = c("GET", "POST", "PUT", "DELETE"),
  Body_Data = NULL
)
}
\arguments{
\item{Request}{The part of the URL that follows after .../API/v3/}

\item{QualBaseUrl}{The part of the URL up to and including .../API/v3/. It's convenient to set this as a global variable.}

\item{Clear_Token}{Gives the option to delete saved credentials to make a correction or remove them from your system's key store.}

\item{Method}{Select from GET, POST, PUT, and DELETE.}

\item{Body_Data}{The data to send with POST and PUT calls.

Body_Data should be formatted like this: 

Body_Data = '
"firstName" = "John",
"lastName" = "Doe",
"email" = "jdoe@email.com",
"unsubscribed" = FALSE
'}
}
\value{
The function returns a data.frame but also outputs a list named QualtricsAPI_Response_Raw which is helpful for debugging and may preserve returned data even if the main function stops with an error.
}
\description{
Use R to interact directly with Qualtrics via the API.
}
\details{
This package takes API endpoints as arguments to give complete flexibility in how you interact with the API, rather than providing more "R-like" functions and arguments.
}
\examples{

Eg. to retrieve data on all surveys in the account:

QualtricsAPI("surveys")


Eg. To create a new mailing list (using POST and sending key value pairs wrapped in single quotes):

QualtricsAPI(
  "mailinglists",
   Method = "POST",
   Body_Data = '
     "category" = "Clients",
     "libraryId" = "[enter your library id here]",
     "name" = "my_new_mailing_list"
              '
            )
                      


}
